fixed warning

+++ tests/cunit/test_async_mpi.c
@@ -134,7 +134,7 @@ int msg_handler(int verbose, int my_rank, int io_rank, int component_count, MPI_
         {
             if (verbose)
                 printf("my_rank %d cmp %d about to call MPI_Irecv comproot[cmp] %d union_comm[cmp] %d\n",
-                       my_rank, cmp, comproot[cmp], union_comm[cmp]);
+                       my_rank, cmp, comproot[cmp], (int)union_comm[cmp]);
             if ((mpierr = MPI_Irecv(&msg, 1, MPI_INT, comproot[cmp], MPI_ANY_TAG,
                                     union_comm[cmp], &req[cmp])))
                 MPIERR(mpierr);
